requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Background processing
celery==5.3.4
redis==5.0.1

# Object storage
minio==7.2.0

# Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3

# Face recognition - PRE-BUILT WHEELS
dlib-binary==19.24.2
face-recognition==1.3.0
click>=6.0